+++ storage/db-core/src/main/java/com/imstargg/storage/db/core/statistics/BrawlerBattleResultStatisticsBaseEntity.java
@@ -25,13 +25,13 @@ abstract class BrawlerBattleResultStatisticsBaseEntity extends BattleStatisticsB
     private boolean duplicateBrawler;
 
     @Column(name = "victory_count", nullable = false)
-    private int victoryCount;
+    private long victoryCount;
 
     @Column(name = "defeat_count", nullable = false)
-    private int defeatCount;
+    private long defeatCount;
 
     @Column(name = "draw_count", nullable = false)
-    private int drawCount;
+    private long drawCount;
 
     protected BrawlerBattleResultStatisticsBaseEntity() {
     }
@@ -50,15 +50,15 @@ abstract class BrawlerBattleResultStatisticsBaseEntity extends BattleStatisticsB
         return duplicateBrawler;
     }
 
-    public int getVictoryCount() {
+    public long getVictoryCount() {
         return victoryCount;
     }
 
-    public int getDefeatCount() {
+    public long getDefeatCount() {
         return defeatCount;
     }
 
-    public int getDrawCount() {
+    public long getDrawCount() {
         return drawCount;
     }
 }
